Refactoring popup createElements
